+++ temba/ivr/tests.py
@@ -4,9 +4,9 @@ from unittest.mock import patch
 from django.urls import reverse
 from django.utils import timezone
 
-from temba.channels.models import ChannelLog
+from temba.channels.models import ChannelConnection, ChannelLog
 from temba.flows.models import FlowSession
-from temba.tests import TembaTest
+from temba.tests import MigrationTest, TembaTest
 from temba.utils.uuid import uuid4
 
 from .models import Call
@@ -78,3 +78,73 @@ class IVRTest(TembaTest):
         response = self.client.get(reverse("mailroom.ivr_handler", args=[self.channel.uuid, "incoming"]))
         self.assertEqual(response.status_code, 404)
         self.assertEqual(response.content, b"this URL should be mapped to a Mailroom instance")
+
+
+class ConvertConnectionsMigrationTest(MigrationTest):
+    app = "ivr"
+    migrate_from = "0020_add_call"
+    migrate_to = "0021_convert_connections"
+
+    def setUpBeforeMigration(self, apps):
+        self.contact1 = self.create_contact("Bob", phone="+123456001")
+        self.contact2 = self.create_contact("Jim", phone="+123456002")
+
+        # create an existing call
+        self.call_kwargs = dict(
+            org=self.org,
+            direction=Call.DIRECTION_IN,
+            status=Call.STATUS_IN_PROGRESS,
+            channel=self.channel,
+            contact=self.contact1,
+            contact_urn=self.contact1.urns.first(),
+            external_id="CA0001",
+            duration=15,
+        )
+        Call.objects.create(**self.call_kwargs)
+
+        self.conn1_kwargs = dict(
+            org=self.org,
+            direction=Call.DIRECTION_IN,
+            status=Call.STATUS_IN_PROGRESS,
+            channel=self.channel,
+            contact=self.contact1,
+            contact_urn=self.contact1.urns.first(),
+            external_id="CA0002",
+            created_on=datetime(2022, 1, 1, 13, 0, 0, 0, timezone.utc),
+            modified_on=datetime(2022, 1, 1, 14, 0, 0, 0, timezone.utc),
+            started_on=datetime(2022, 1, 1, 15, 0, 0, 0, timezone.utc),
+            ended_on=None,
+            duration=None,
+            error_reason=None,
+            error_count=0,
+            next_attempt=None,
+        )
+        ChannelConnection.objects.create(**self.conn1_kwargs)
+
+        self.conn2_kwargs = dict(
+            org=self.org,
+            direction=Call.DIRECTION_OUT,
+            status=Call.STATUS_ERRORED,
+            channel=self.channel,
+            contact=self.contact2,
+            contact_urn=self.contact2.urns.first(),
+            external_id="",
+            created_on=datetime(2022, 1, 2, 13, 0, 0, 0, timezone.utc),
+            modified_on=datetime(2022, 1, 2, 14, 0, 0, 0, timezone.utc),
+            started_on=datetime(2022, 1, 2, 15, 0, 0, 0, timezone.utc),
+            ended_on=datetime(2022, 1, 2, 16, 0, 0, 0, timezone.utc),
+            duration=15,
+            error_reason=Call.ERROR_BUSY,
+            error_count=1,
+            next_attempt=datetime(2022, 1, 2, 16, 0, 0, 0, timezone.utc),
+        )
+        ChannelConnection.objects.create(**self.conn2_kwargs)
+
+    def test_migration(self):
+        # all connections gone
+        self.assertEqual(0, ChannelConnection.objects.count())
+        self.assertEqual(3, Call.objects.count())
+
+        self.assertTrue(Call.objects.filter(**self.call_kwargs).exists())
+        self.assertTrue(Call.objects.filter(**self.conn1_kwargs).exists())
+        self.assertTrue(Call.objects.filter(**self.conn2_kwargs).exists())